requirements.txt
smbus2
influxdb
numpy
PyYAML
RPi.GPIO
